+++ src/utils/ff_send_info.h
@@ -13,7 +13,7 @@
  *        when and what diagonal to send to the Workers
  */
 struct SendInfo {
-    SendInfo(const u16 num_workers, const u16 base_lenght)
+    SendInfo(const u64 num_workers, const u64 base_lenght)
     : diag_length(base_lenght), num_workers(num_workers) {PrepareNextDiagonal();}
 
     /**
@@ -28,10 +28,10 @@ struct SendInfo {
     }
 
     // PARAMETERS
-    u16 diag{0};                // The current diagonal.
-    u16 diag_length{0};         // Number of elements of the curr. diagonal.
-    u16 num_workers{0};         // Number of elements per Worker.
-    u16 computed_elements{0};   // Number of computed elements during a cycle
+    u64 diag{0};                // The current diagonal.
+    u64 diag_length{0};         // Number of elements of the curr. diagonal.
+    u64 num_workers{0};         // Number of elements per Worker.
+    u64 computed_elements{0};   // Number of computed elements during a cycle
     bool send_tasks{false};     // Tells if we have to send tasks to Workers.
 };
